move authentication startup code into its own self-contained, more easily reusable class
